+++ CHANGELOG.md
@@ -1,1 +1,10 @@
 # This file contains the Change logs.
+All notable changes to this project will be documented in this file.
+
+# [1.1.0] - 2023-05-23
+### Enhanced
+- Implemented signing verification process for cdn file content
+
+## [1.0.0] - 2023-18-04
+### Added
+- Initial public release